Segreggated component TestGeneration and update page
